+++ include/sqare_matrix.h
@@ -13,7 +13,7 @@ namespace numerical{
     class SquareMatrix{
     public:
         // Creates matrix full of zeros
-        explicit SquareMatrix(int64_t size);
+        explicit SquareMatrix(int32_t size);
 
         // Loads matrix from file
         explicit SquareMatrix(const std::string& filename);
@@ -23,11 +23,11 @@ namespace numerical{
         void Write(const std::string& filename);
 
         // Returns elements of the matrix
-        double& operator() (int64_t i, int64_t j);
-        const double& operator () (int64_t i, int64_t j) const;
+        double& operator() (int32_t i, int32_t j);
+        const double& operator () (int32_t i, int32_t j) const;
 
         // Return row of the matrix
-        double*& operator () (int64_t i) const;
+        double*& operator () (int32_t i) const;
 
         // Copy operator
         SquareMatrix& operator =(const SquareMatrix& other);
@@ -38,11 +38,11 @@ namespace numerical{
         SquareMatrix& operator -=(const SquareMatrix& other);
 
         // returns matrix dimension
-        int64_t Size() const;
+        int32_t Size() const;
     private:
-        static double** _allocate_matrix(int64_t size);
+        static double** _allocate_matrix(int32_t size);
 
-        int64_t size;
+        int32_t size;
         double **matrix;
     };
 }
